Initial commit
Software is currently in a working state. Code needs to be manually edited to change what historical data is fetched, it may be better for this to be handled via a command-line argument or configuration file. We could also be smarter about how we fill in data if it exists - currently if data exists in the database, we wipe it out first. Instead we should hunt for gaps between what we have and what we could have, and synchronize the difference.